Not needed
git-svn-id: svn://svn.code.sf.net/p/tass64/code/trunk@2959 111d7451-93fa-4e81-90d4-8039f76616e9

+++ instruction.c
@@ -661,7 +661,7 @@ MUST_CHECK Error *instruction(int prm, unsigned int w, Funcargs *vals, linepos_t
                     }
                 }
             } else if (!invalid) { /* short */
-                if (((uint16_t)(current_address->l_address + 1 + ln) & 0xff00) != (oadr & 0xff00)) {
+                if (((current_address->l_address + 1 + ln) & 0xff00) != (oadr & 0xff00)) {
                     int diff = (int8_t)oadr;
                     if (diff >= 0) diff++;
                     if (!allowslowbranch) err_msg2(ERROR__BRANCH_CROSS, &diff, epoint2);
